# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/retail_eggtv": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "이용권 목록",
                "operationId": "8a35aa316ece75cfecc80e76ff42cc0f",
                "parameters": [
                    {
                        "name": "operatorId",
                        "in": "query",
                        "description": "운영자 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "팀 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "정상 처리",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "eggtv_ticket_id": {
                                                    "description": "이용권 ID",
                                                    "type": "string",
                                                    "example": "109"
                                                },
                                                "eggtv_ticket_name": {
                                                    "description": "이용권 이름",
                                                    "type": "string",
                                                    "example": "라이트 이용권"
                                                },
                                                "eggtv_ticket_type": {
                                                    "description": "이용권 유형(0=SUBSCRIPTION=기간 이용권,1=TIER_UPGRADE=구독 등급 업그레이드)",
                                                    "type": "string",
                                                    "example": "0"
                                                },
                                                "subscription_tier": {
                                                    "description": "구독 티어(1=LITE=라이트,2=STANDARD=스탠다드,3=PLUS=플러스,4=PREMIUM=프리미엄)",
                                                    "type": "string",
                                                    "example": "1"
                                                },
                                                "has_fixed_end_date": {
                                                    "description": "이용권의 종료 일자가 명시적으로 정해져 있는지 여부",
                                                    "type": "string",
                                                    "example": "0"
                                                },
                                                "subscription_days": {
                                                    "description": "구독 기간 OR 업그레이드 적용 기간(단위: 일)",
                                                    "type": "string",
                                                    "example": "20"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "조회 실패",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "NOT_FOUND_MEMBER_INFO"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "회원정보를 찾을 수 없습니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "이용권 추가",
                "operationId": "dded7822d9240a168a01184d79b3b0d9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "operatorId": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "teamId": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "에그티비 7일 이용권"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "tier": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "hasEndDate": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "days": {
                                        "type": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "정상 처리",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "eggtv_ticket_id": {
                                                    "description": "이용권 ID",
                                                    "type": "string",
                                                    "example": "109"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "NOT_FOUND_MEMBER_INFO"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "회원정보를 찾을 수 없습니다."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "조회 실패",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "DUPLICATED_EGGTV_TICKET_CONDITION || MISSING_REQUIRED_DATA"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "이미 등록된 에그TV 이용권 조건입니다. || 데이터를 입력해주세요."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Ticket",
            "description": "EGG TV 이용권 관련 API. 이용권 조회, 발행, 코드 발급 등을 포함합니다."
        }
    ]
}